### 2023-03-20 #5281 @editorialbot
[PRE REVIEW]: LabCD: An annotation tool for remote sensing change detection
**Submitting author:** @geoyee (<a href="http://orcid.org/0000-0002-4901-5866">YiZhou Chen</a>)
**Repository:** https://github.com/geoyee/LabCD
**Branch with paper.md** (empty if default branch): docs
**Version:** 0.3
