+++ paper/paper.bib
@@ -99,6 +99,7 @@
   pages={611--622},
   year={1999},
   publisher={ASPRS AMERICAN SOCIETY FOR PHOTOGRAMMETRY AND}
+  doi={10.1002/9780470979587.ch22}
 }
 
 @article{chavez1988improved,
