+++ paper/paper.bib
@@ -98,7 +98,7 @@
   volume={65},
   pages={611--622},
   year={1999},
-  publisher={ASPRS AMERICAN SOCIETY FOR PHOTOGRAMMETRY AND}
+  publisher={ASPRS AMERICAN SOCIETY FOR PHOTOGRAMMETRY AND},
   doi={10.1002/9780470979587.ch22}
 }
 
