### 2023-04-05 #5344 @editorialbot
[REVIEW]: LabCD: An annotation tool for remote sensing change detection
**Submitting author:** @geoyee (<a href="http://orcid.org/0000-0002-4901-5866">YiZhou Chen</a>)
**Repository:** https://github.com/geoyee/LabCD
**Branch with paper.md** (empty if default branch): docs
**Version:** 0.3
