### 2023-04-21 #5344 @azarakhshrafiee
I cannot find "Automatically calculate optimum index factor (OIF)" functionality in the tool, as it is mentioned in the paper.
