### 2023-04-24 #5344 @geoyee
> I cannot find "Automatically calculate optimum index factor (OIF)" functionality in the tool, as it is mentioned in the paper.

if you load hsi image, OIF will be calculated automatically and canvas will load optimal bands. in [here](https://github.com/geoyee/LabCD/blob/master/labcd/utils/imgpress.cpp#L394).

``` cpp
...
if (bandCount < 3)  // only load first channel
{
	bandList.append(poDataset->GetRasterBand(1));
	bandList.append(poDataset->GetRasterBand(1));
	bandList.append(poDataset->GetRasterBand(1));
}
else if (bandCount == 3)  // load directly if channels is 3
{
	bandList.append(poDataset->GetRasterBand(1));
	bandList.append(poDataset->GetRasterBand(2));
	bandList.append(poDataset->GetRasterBand(3));
}
else  // hsi image
{
	// only be calculated once
	if (HSI_RGB_LOAD.empty() || bandCount != NOW_BAND_COUNT)
		HSI_RGB_LOAD = ImagePress::calcOIF(imgPath);
	// load optimal band
	bandList.append(poDataset->GetRasterBand(HSI_RGB_LOAD[0]));
	bandList.append(poDataset->GetRasterBand(HSI_RGB_LOAD[1]));
	bandList.append(poDataset->GetRasterBand(HSI_RGB_LOAD[2]));
}
```
