### 2023-05-25 #5344 @hugoledoux
@azarakhshrafiee the goal is that all points are checked in the list, so if you didn't check is this because you haven't had time yet to check, or because the submission doesn't fulfil those criteria? If so, could you please elaborate so that @geoyee can take action on those?
