### 2023-05-28 #5344 @geoyee
> @hugoledoux : Sure! I have now updated the checklist. @geoyee : I missed a summary section in your paper.

@hugoledoux Ok, I'll add it by Monday
