### 2023-05-28 #5344 @geoyee
> Dear @geoyee,
>
> The paper is well structured and organized, and the tool is timely for the remote sensing community using machine learning techniques. I appreciate the efforts of the author in building this tool. I have the following questions/suggestions.
>
> 1. I can only see a non-English version of the Windows binary. An English version of binary would be helpful for more reseacrhers.
> 2. The introduction section could be improved in terms of language and technical details. It is difficult to understand because of the lack of connectivity between sentences. A grammar check for the entire manuscript is highly recommended.
> 3. **Line 22-23:** Please verify the abbreviation of GIS in the statement of need section. A widely used abbreviation GIS is : geographic information system.
> 4. **Line 23-24:** `"By the way, they are not efficient for annotation for remote sensing change detection`."
>    Most researchers in the remote sensing domain have used ArcGIS, QGIS, and ENVI. Therefore, the above statement is a very strong statement (subjective opinion) without any references. I suggest the authors rephrase/justify and cite proper references.
> 5. **Line 25-26:** "`For example, ArcGIS and QGIS can't create two synchronized windows,...`"
>    I have to disagree with the about statement because QGIS does support synchronous windows. In fact, it supports multiple synchronous windows with more flexible functionalities like independent scale, recentering, etc. I've included a sample screenshot for you to look over. Further, the proposed functionalities like change vector analysis ([whitebox tools](https://www.whiteboxgeo.com/manual/wbt_book/available_tools/image_processing_tools.html#ChangeVectorAnalysis)) and optimum index factor (GRSS GIS [oif ](https://grass.osgeo.org/grass82/manuals/i.oif.html)function ), are readily available with QGIS plugins/extensions/toolboxes. Therefore, the authors need to provide sufficient details about the additional contributions of the tool to the scientific community.
>
> ![QGIS_multi_window](https://user-images.githubusercontent.com/50410669/241239024-df05e82d-ef60-4136-85d9-35208b003ed5.png)
>
> 7. I suggest keeping a title or description to the table and citing it appropriately in the text.

@Narayana-Rao It can be set to switch to the English version, and I will try to solve other problems before Monday. Thanks for your advice
